requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.0
langdetect==1.0.9
openai==1.3.5
anthropic==0.7.7
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.2.3
selenium==4.15.2
webdriver-manager==4.0.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
language-tool-python==2.7.1
polyglot==16.7.4
fasttext==0.9.2
pycld3==0.22
google-cloud-translate==3.12.1
deepl==1.16.1
